Fix migration 002 for tests on 2.6
Migration 002 was written before the migration was fixed by clearing
the mappers. It was attempting to bootstrap a sqlalchemy connection
itself which would fail under tests on 2.6 because it couldn't find
the config file. Since this is no longer needed, execute the query
via the migration_engine that is already setup.